Added collector-bitmovin-player and collector-exoplayer modules
