+++ analyticsexample/build.gradle
@@ -36,6 +36,8 @@ dependencies {
 
     //Compile from source
     compile project(path: ':collector')
+    compile project(path: ':collector-exoplayer')
+    implementation 'com.google.android.exoplayer:exoplayer:2.8.+'
 
     // Pull from Maven Repo
 //    compile 'com.bitmovin.analytics:collector:1.3.8'
