+++ bitmovinanalyticsexample/build.gradle
@@ -38,7 +38,10 @@ dependencies {
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
 
     compile project(path: ':collector')
+    compile project(path: ':collector-bitmovin-player')
+    implementation 'com.bitmovin.player:playercore:2.13.0'
 
 //    compile 'com.bitmovin.analytics:collector:1.3.8'
 
 }
+
