+++ collector-bitmovin-player/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'com.android.library'
 apply plugin: 'com.jfrog.artifactory'
 apply plugin: 'maven-publish'
 
-def packageName = 'com.bitmovin.analytics'
+def packageName = 'com.bitmovin.analytics-bitmovin-player'
 def libraryVersion = '1.3.8'
 
 android {
@@ -37,6 +37,9 @@ dependencies {
 
     compile 'com.squareup.okhttp3:okhttp:3.9.1'
     compile 'com.google.code.gson:gson:2.8.2'
+    implementation project(path: ':collector')
+    implementation 'com.bitmovin.player:playercore:2.13.0'
+
 }
 
 publishing {
