+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/adapters/BitmovinSdkAdapter.java
