+++ collector/src/main/java/com/bitmovin/analytics/utils/Util.java
@@ -6,7 +6,6 @@ import android.content.res.Resources;
 import android.provider.Settings;
 
 import com.bitmovin.analytics.BuildConfig;
-import com.google.android.exoplayer2.Player;
 
 import java.util.UUID;
 
@@ -27,31 +26,6 @@ public class Util {
 
     public static String getVersion() {
         return BuildConfig.VERSION_NAME;
-    }
-
-    public static String exoStateToString(int state) {
-        switch (state) {
-            case Player.STATE_IDLE:
-                return "Idle";
-            case Player.STATE_BUFFERING:
-                return "Buffering";
-            case Player.STATE_READY:
-                return "Ready";
-            case Player.STATE_ENDED:
-                return "Ended";
-            default:
-                return "Unknown PlayerState";
-        }
-    }
-
-    public static String getUserAgent(Context context) {
-        ApplicationInfo applicationInfo = context.getApplicationInfo();
-        int stringId = applicationInfo.labelRes;
-        String applicationName = "Unknown";
-        if (stringId == 0 && applicationInfo.nonLocalizedLabel != null) {
-            applicationInfo.nonLocalizedLabel.toString();
-        }
-        return com.google.android.exoplayer2.util.Util.getUserAgent(context, applicationName);
     }
 
     public static long getTimeStamp() {
